+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Updated phpcs ruleset to PSR-12
 - is_null() changed to === null
 - Using identity comparison instead of equality comparison
+- Removed checks in ifs that var is (not) null before empty(var) function (empty checks that itself)
 
 #### Fixed
 - Fixed wrong dictionary name in post.php